move the data part of the script to the data
